orsr/requirements.txt
basicsr==1.4.2
numpy==2.2.0
Pillow==11.0.0
pytest==8.3.4
opencv-python~=4.10.0.84
pytorch_msssim==1.0.0
PyYAML==6.0.2
realesrgan==0.3.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
